+++ app/src/main/java/com/rma/expensetracker/data/models/mock/SavingsAccount.kt
@@ -5,6 +5,6 @@ import androidx.compose.ui.graphics.vector.ImageVector
 data class SavingsAccount(
     val id: String,
     val title: String,
-    val balance: Float,
+        val balance: Double,
     val icon: ImageVector? = null  //TODO
 )
